fix: windows type
