+++ plat_win/storage_types.go
@@ -1,10 +1,5 @@
 package plat_win
 
-import (
-	"golang.org/x/sys/windows"
-	"unsafe"
-)
-
 type PartitionStyle uint32
 
 const (
@@ -63,47 +58,6 @@ type DISK_EXTENT struct {
 type VOLUME_DISK_EXTENTS struct {
 	NumberOfDiskExtents uint32
 	Extents             [1]DISK_EXTENT
-}
-
-type PARTITION_INFORMATION_MBR struct {
-	PartitionType       byte
-	BootIndicator       bool
-	RecognizedPartition bool
-	HiddenSectors       uint32
-	PartitionId         windows.GUID
-}
-
-type PARTITION_INFORMATION_GPT struct {
-	PartitionType windows.GUID
-	PartitionId   windows.GUID
-	Attributes    uint64
-	Name          [36]uint16
-}
-
-type PARTITION_INFORMATION_EX struct {
-	PartitionStyle   PartitionStyle
-	StartingOffset   int64
-	PartitionLength  int64
-	PartitionNumber  int32
-	RewritePartition bool
-	Rev01            bool
-	Rev02            bool
-	Rev03            bool
-	PartitionInfo    [112]byte
-}
-
-func (p *PARTITION_INFORMATION_EX) GetMbr() *PARTITION_INFORMATION_MBR {
-	if p.PartitionStyle == PartitionStyleGpt {
-		return (*PARTITION_INFORMATION_MBR)(unsafe.Pointer(&p.PartitionInfo[0]))
-	}
-	return nil
-}
-
-func (p *PARTITION_INFORMATION_EX) GetGpt() *PARTITION_INFORMATION_GPT {
-	if p.PartitionStyle == PartitionStyleGpt {
-		return (*PARTITION_INFORMATION_GPT)(unsafe.Pointer(&p.PartitionInfo[0]))
-	}
-	return nil
 }
 
 type STORAGE_BUS_TYPE byte
